// src/app/api/models/LoginOutputDTOResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LoginOutputDTO } from './LoginOutputDTO';
export type LoginOutputDTOResponseDTO = {
    message?: string | null;
    status?: number | null;
    data?: LoginOutputDTO;
    count?: number | null;
};